requirements.txt
fastapi
uvicorn
python-dotenv
pdfkit
jinja2
passlib[bcrypt]
openai>=1.0.0



